+++ flask_app/mocks/constants.py
@@ -1,0 +1,9 @@
+"""
+This file contains constant prefixed data.
+"""
+
+DISMISSED_LESSON_NAMES = [
+    "Uczniowie zwolnieni do domu",
+    "Uczniowie przychodzą później",
+    "Okienko dla uczniów",
+]
